requirements.txt
Django==6.0.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
